# apps/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/discount-coupons": {
      "get": {
        "description": "Renvoie tous les coupons de réduction de l'organisation du marchand authentifié.",
        "operationId": "DiscountCouponsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Liste des coupons",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DiscountCouponResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Clé API invalide ou manquante"
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Lister les coupons",
        "tags": ["Coupons de réduction"]
      },
      "post": {
        "description": "Crée un coupon dans votre organisation. Le code sera automatiquement mis en majuscules.",
        "operationId": "DiscountCouponsController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Coupon créé avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscountCouponResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Données invalides ou code déjà utilisé"
          },
          "401": {
            "description": "Clé API invalide ou manquante"
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Créer un coupon",
        "tags": ["Coupons de réduction"]
      }
    },
    "/discount-coupons/{id}/performance": {
      "get": {
        "description": "Renvoie les statistiques d'utilisation et l'impact sur le chiffre d'affaires pour un coupon donné.",
        "operationId": "DiscountCouponsController_getPerformance",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID du coupon",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Indicateurs de performance",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "total_uses": {
                      "type": "number",
                      "example": 45
                    },
                    "total_discount_amount": {
                      "type": "number",
                      "example": 25000
                    },
                    "total_revenue": {
                      "type": "number",
                      "example": 150000
                    },
                    "average_order_value": {
                      "type": "number",
                      "example": 3333.33
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Clé API invalide ou manquante"
          },
          "404": {
            "description": "Coupon introuvable ou accès refusé"
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Indicateurs de performance du coupon",
        "tags": ["Coupons de réduction"]
      }
    },
    "/discount-coupons/{id}": {
      "get": {
        "description": "Renvoie un coupon. Réponse 404 s'il n'existe pas ou n'est pas accessible avec cette clé API.",
        "operationId": "DiscountCouponsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID du coupon",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Détails du coupon",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscountCouponResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Clé API invalide ou manquante"
          },
          "404": {
            "description": "Coupon introuvable ou accès refusé"
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Obtenir un coupon par ID",
        "tags": ["Coupons de réduction"]
      }
    },
...
    "/payment-links": {
      "post": {
        "description": "Crée un lien de paiement partageable : produit (lien produit) ou montant fixe (lien instantané). Les liens produit référencent un produit et un prix optionnel ; les liens instantanés ont un montant fixe.",
        "operationId": "PaymentLinksController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["link_type", "title", "currency_code"],
                "properties": {
                  "link_type": {
                    "type": "string",
                    "enum": ["product", "instant"]
                  },
                  "title": {
                    "type": "string"
                  },
                  "currency_code": {
                    "type": "string",
                    "enum": ["XOF", "USD", "EUR"]
                  },
                  "description": {
                    "type": "string"
                  },
                  "amount": {
                    "type": "number"
                  },
                  "product_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "price_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "allow_coupon_code": {
                    "type": "boolean"
                  },
                  "allow_quantity": {
                    "type": "boolean"
                  },
                  "require_billing_address": {
                    "type": "boolean"
                  },
                  "expires_at": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "success_url": {
                    "type": "string",
                    "format": "uri"
                  },
                  "cancel_url": {
                    "type": "string",
                    "format": "uri"
                  },
                  "metadata": {
                    "type": "object",
                    "additionalProperties": true
                  }
                }
              },
              "examples": {
                "instant": {
                  "summary": "Lien instantané (montant fixe)",
                  "value": {
                    "link_type": "instant",
                    "title": "Invoice INV-001",
                    "currency_code": "XOF",
                    "amount": 15000
                  }
                },
                "product": {
                  "summary": "Lien produit",
                  "value": {
                    "link_type": "product",
                    "title": "Premium plan",
                    "currency_code": "XOF",
                    "product_id": "123e4567-e89b-12d3-a456-426614174000"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Lien de paiement créé avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentLinkResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Entrée invalide ou erreur de validation"
          },
          "401": {
            "description": "Clé API invalide ou manquante"
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Créer un lien de paiement",
        "tags": ["Liens de paiement"]
      },
      "get": {
        "description": "Renvoie tous les liens de l'organisation avec pagination et filtres optionnels (type, actif).",
        "operationId": "PaymentLinksController_findAll",
        "parameters": [
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Décalage pour la pagination",
            "schema": {
              "example": 0,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Nombre maximal de résultats",
            "schema": {
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "Filtrer par statut actif",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "linkType",
            "required": false,
            "in": "query",
            "description": "Filtrer par type de lien",
            "schema": {
              "enum": ["product", "instant"],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste paginée de liens de paiement",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/PaymentLinkResponseDto"
                      }
                    },
                    "total": {
                      "type": "number"
                    },
                    "limit": {
                      "type": "number"
                    },
                    "offset": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Clé API invalide ou manquante"
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Lister les liens de paiement",
        "tags": ["Liens de paiement"]
      }
    },
    "/payment-links/{id}": {
      "get": {
        "description": "Renvoie le détail d’un lien (URL, configuration et statut).",
        "operationId": "PaymentLinksController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID du lien de paiement",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Détails du lien",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentLinkResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Clé API invalide ou manquante"
          },
          "404": {
            "description": "Lien introuvable ou accès refusé"
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Obtenir un lien de paiement par ID",
        "tags": ["Liens de paiement"]
      }
    },
...
  },
  "info": {
    "title": "API lomi.",
    "description": "API de traitement des paiements pour les entreprises d'Afrique de l'Ouest francophone.",
    "version": "1.1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://api.lomi.africa",
      "description": "Live"
    },
    {
      "url": "https://sandbox.api.lomi.africa",
      "description": "Test"
    }
  ],
  "components": {
    "securitySchemes": {
      "api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-KEY"
      }
    },
    "schemas": {
...
      "DiscountCouponResponseDto": {
        "type": "object",
        "properties": {
          "coupon_id": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "description": "Unique coupon identifier"
          },
          "organization_id": {
            "type": "string",
            "example": "789e0123-e89b-12d3-a456-426614174000",
            "description": "Organization ID the coupon belongs to"
          },
          "code": {
            "type": "string",
            "example": "SAVE20",
            "description": "Unique coupon code"
          },
          "discount_type": {
            "type": "string",
            "example": "percentage",
            "description": "Type of discount",
            "enum": ["percentage", "fixed"]
          },
          "discount_percentage": {
            "type": "number",
            "example": 20,
            "description": "Discount percentage (if discount_type is percentage)",
            "nullable": true
          },
          "discount_fixed_amount": {
            "type": "number",
            "example": 1000,
            "description": "Fixed discount amount (if discount_type is fixed)",
            "nullable": true
          },
          "customer_type": {
            "type": "string",
            "example": "all",
            "description": "Customer type this coupon applies to",
            "enum": ["all", "new", "existing"]
          },
          "is_active": {
            "type": "boolean",
            "example": true,
            "description": "Whether the coupon is currently active"
          },
          "max_uses": {
            "type": "number",
            "example": 100,
            "description": "Maximum number of times this coupon can be used",
            "nullable": true
          },
          "max_quantity_per_use": {
            "type": "number",
            "example": 5,
            "description": "Maximum quantity allowed per use",
            "nullable": true
          },
          "current_uses": {
            "type": "number",
            "example": 25,
            "description": "Current number of times this coupon has been used"
          },
          "usage_frequency_limit": {
            "type": "string",
            "example": "total",
            "description": "Usage frequency limit type",
            "enum": ["total", "per_customer", "per_customer_per_product"]
          },
          "usage_limit_value": {
            "type": "number",
            "example": 1,
            "description": "Usage limit value (for per_customer limits)",
            "nullable": true
          },
          "expires_at": {
            "type": "string",
            "example": "2024-12-31T23:59:59Z",
            "description": "When the coupon expires",
            "nullable": true
          },
          "valid_from": {
            "type": "string",
            "example": "2024-01-01T00:00:00Z",
            "description": "When the coupon becomes valid",
            "nullable": true
          },
          "description": {
            "type": "string",
            "example": "20% off all products",
            "description": "Coupon description",
            "nullable": true
          },
          "scope_type": {
            "type": "string",
            "example": "organization_wide",
            "description": "Scope of the coupon",
            "enum": [
              "organization_wide",
              "specific_products",
              "specific_prices"
            ]
          },
          "applies_to_product_types": {
            "example": ["one_time", "recurring"],
            "description": "Product types this coupon applies to",
            "nullable": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "environment": {
            "type": "string",
            "example": "live",
            "description": "Environment (test or live)",
            "enum": ["test", "live"]
          },
          "created_at": {
            "type": "string",
            "example": "2024-01-15T10:30:00Z",
            "description": "When the coupon was created"
          },
          "updated_at": {
            "type": "string",
            "example": "2024-01-15T10:30:00Z",
            "description": "When the coupon was last updated"
          }
        },
        "required": [
          "coupon_id",
          "organization_id",
          "code",
          "discount_type",
          "discount_percentage",
          "discount_fixed_amount",
          "customer_type",
          "is_active",
          "max_uses",
          "max_quantity_per_use",
          "current_uses",
          "usage_frequency_limit",
          "usage_limit_value",
          "expires_at",
          "valid_from",
          "description",
          "scope_type",
          "applies_to_product_types",
          "environment",
          "created_at",
          "updated_at"
        ]
      },
...
      "PaymentLinkResponseDto": {
        "type": "object",
        "properties": {
          "allow_coupon_code": {
            "type": "boolean",
            "example": true
          },
          "allow_quantity": {
            "type": "boolean",
            "example": true
          },
          "amount": {
            "type": "number",
            "example": 123
          },
          "cancel_url": {
            "type": "string",
            "example": "string"
          },
          "created_at": {
            "type": "string",
            "example": "string"
          },
          "created_by": {
            "type": "string",
            "example": "string"
          },
          "currency_code": {
            "type": "string",
            "example": "string"
          },
          "description": {
            "type": "string",
            "example": "string"
          },
          "environment": {
            "type": "string",
            "example": "string"
          },
          "expires_at": {
            "type": "string",
            "example": "string"
          },
          "is_active": {
            "type": "boolean",
            "example": true
          },
          "link_id": {
            "type": "string",
            "example": "string"
          },
          "link_type": {
            "type": "string",
            "example": "string"
          },
          "metadata": {
            "type": "object",
            "example": {}
          },
          "organization_id": {
            "type": "string",
            "example": "string"
          },
          "price_id": {
            "type": "string",
            "example": "string"
          },
          "product_id": {
            "type": "string",
            "example": "string"
          },
          "quantity": {
            "type": "number",
            "example": 123
          },
          "require_billing_address": {
            "type": "boolean",
            "example": true
          },
          "success_url": {
            "type": "string",
            "example": "string"
          },
          "title": {
            "type": "string",
            "example": "string"
          },
          "updated_at": {
            "type": "string",
            "example": "string"
          },
          "url": {
            "type": "string",
            "example": "string"
          }
        },
        "required": [
          "allow_coupon_code",
          "allow_quantity",
          "amount",
          "cancel_url",
          "created_at",
          "created_by",
          "currency_code",
          "description",
          "environment",
          "expires_at",
          "is_active",
          "link_id",
          "link_type",
          "metadata",
          "organization_id",
          "price_id",
          "product_id",
          "quantity",
          "require_billing_address",
          "success_url",
          "title",
          "updated_at",
          "url"
        ]
      },
...
    }
  }
}